Support ignoring files within the native isort configuration.
This commit adds proper support for ignoring files in the isort configuration,
without duplicating that configuration in pytest_isort's 'isort_ignore'
setting.

Note:
- For isort>=5, skipping files in the native isort configuration was
  unsupported, you were forced to ignore files using the 'isort_ignore'
  setting.
- For isort<5, skipping files in the native isort configuration was actually
  supported, but the file was reported as PASS and not as SKIPPED.

With this change, the isort configuration is respected, and skipped files are
reported as such. Possibly, the 'isort_ignore' configuration parameter could
be deprecated in the future.